Revert "玩家捡+使用武器"
