+++ Assets/Scenes/1.unity
@@ -146,10 +146,6 @@ PrefabInstance:
       propertyPath: m_Color.r
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 8534018703369305741, guid: 0502ff3ce5c2047b8a4b7cbe649c8058, type: 3}
-      propertyPath: attackCooldown
-      value: 0.2
-      objectReference: {fileID: 0}
     - target: {fileID: 8534018703369305743, guid: 0502ff3ce5c2047b8a4b7cbe649c8058, type: 3}
       propertyPath: m_RootOrder
       value: 2
@@ -1364,68 +1360,3 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2096438502}
   m_CullTransparentMesh: 1
---- !u!1001 &8031145810198628498
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_RootOrder
-      value: 9
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 1.1296067
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalScale.y
-      value: 0.8996
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0.37
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -3.59
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0.10464853
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1203859684514259510, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6780747909390881956, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
-      propertyPath: m_Name
-      value: PlayerWeapon
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 45b9608e0ebf72c4091d8711fbe82ad6, type: 3}
