--- Assets/resources/Items/PlayerWeapon.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d8b39b4ee2a1a174989295762b04f035, type: 3}
-  m_Name: PlayerWeapon
-  m_EditorClassIdentifier: 
-  itemName: PlayerWeapon
-  icon: {fileID: 0}
-  itemType: 0
