--- Assets/resources/Items/PlayerWeapon.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eff75786a295dcd48a2ff8dede2a54a6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
