--- Assets/resources/Items/PlayerWeapon.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 45b9608e0ebf72c4091d8711fbe82ad6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
