+++ Assets/resources/weapon.meta
