--- Assets/scripts/InWorldItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 83672a163eccd534b977893bbce38c65
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
