--- Assets/scripts/Item.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-
-public enum ItemType { Weapon, Key, Consumable }
-
-[CreateAssetMenu(fileName = "New Item", menuName = "Inventory/Item")]
-public class Item : ScriptableObject
-{
-    public string itemName;
-    public Sprite icon;
-    public ItemType itemType;
-}
