--- Assets/scripts/Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d8b39b4ee2a1a174989295762b04f035
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
