--- Assets/scripts/PlayerInventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 341c92597a280d848ab9d1c0a0519af3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
